feat(task 6): add combine reducers
