+++ src/reducers/actionCreator.js
@@ -1,4 +1,4 @@
+import {ADD, EDIT} from "../constants";
 
-
-export const createProduct = (value)=>({type: 'ADD',value})
-export const editProduct = (value)=>({type: 'EDIT',value})
+export const createProduct = (value)=>({type: ADD,value})
+export const editProduct = (value)=>({type: EDIT,value})
